refactor(cli): move ui package to internal/ui
Relocate the shared ui package out of internal/cli so non-cli packages (e.g.
internal/packages, which prints install progress) can render with the same
styled primitives without a cli->packages layering inversion. No behaviour
change; only the import path moves.
